error at uppercase and lowercase in mongo query

+++ src/controllers/express/api/ethereum/upload.js
@@ -2,12 +2,15 @@ const ethers = require('ethers');
 const config = require('../../../../config');
 const response_express = require(config.library_dir + '/response').response_express;
 const User = require(config.models_dir + '/mongo/user');
+const History = require(config.models_dir + '/mongo/history');
 const Music = require(config.models_dir + '/mongo/music');
 // const Notification = require(config.models_dir + '/mongo/notification');
 
 module.exports = async (req, res) => {
     try {
         const user = await User.findById(req.token_info._id)
+        .lean()
+        .select('privateKey socketID')
         if(!user){
             return response_express.exception(res, "User not exist!")
         }
@@ -17,30 +20,31 @@ module.exports = async (req, res) => {
         let wallet = new ethers.Wallet(privateKey, config.provider);
         let contractWithSigner = new ethers.Contract(config.userBehaviorAddress, config.userBehaviorABI, wallet)
         contractWithSigner.uploadFile(req.body.ether.hash, req.body.ether.price, 2, music._id.toString())
-        .then(async (tx) => {
-            
-            if(!tx){
+        .then(async (transaction) => {
+            if(!transaction){
+                music.deleteOne()
                 return response_express.exception(res, "Transaction is null ")
             }
-            
-            const receipt = await tx.wait()
+            const receipt = await transaction.wait()
+            if(receipt.status !== 1){
+                music.deleteOne()
+                return response_express.exception(res, "Receipt not exist!");
+            }
             const numb = parseInt(receipt.logs[0].data.slice(130), 16) // Get Id from event uploadFile
-            // const dataNotification = {
-            //     userID: req.token_info._id,
-            //     content: req.body.server.name + " - Copyright registration is successful",
-            //     link: `song/${music._id}`,
-            // }
-            const promises = [
-                Music.updateOne({ _id: music._id }, { $set: { idSolidity: numb } }).exec(),
-                // Notification.create(dataNotification)
-            ]
-            
-            await Promise.all(promises)
-            
-            return response_express.success(res, tx.hash)
+            Music.updateOne({ _id: music._id }, { $set: { idSolidity: numb } }).exec()
+            response_express.success(res, receipt.transactionHash)
+            const historyData = {
+                senderID: req.token_info._id,
+                songID: music._id,
+                contentSender: `Đăng ký bản quyền tác phẩm \"${req.body.server.name}\" thành công.`,
+                type: 1,
+                money: req.body.ether.price,
+                songImage: req.body.server.image
+            }
+            const newHistory = await History.create(historyData)
+            socket.to(user.socketID).emit('notification', newHistory);
         })
         .catch( (error) => {
-            music.deleteOne()
             return response_express.exception(res, "Field to execute transaction: " + error)
         })
     } catch (error) {
